+++ proto/control-api/CHANGELOG.md
@@ -15,8 +15,13 @@ All user visible changes to this project will be documented in this file. This p
 
 - Abstractions:
     - `ControlApi` and `CallbackApi` ([#55]).
+- gRPC:
+    - `ControlApi` and `CallbackApi` implementations ([#56]).
+- Cargo features:
+    - `client` and `server` ([#56]).
 
 [#55]: /../../pull/55
+[#56]: /../../pull/56